Fixed LVALUE/RVALUE field conflict
- Fixed the issue where field processing was assuming the value was only an LVALUE, leading to bad pointer referencing and crashes.
- Added ability to label a type as reference type in the methods file, indicating the value should be returned as an LVALUE instead of an RVALUE.  This will only work if the field and/or method are defined to handle it.
- Added ARRAY type.
- Added ability to defined LIST and ARRAY types in methods as having CONSTANT elements: eg. list(constant mobile)
- Added FLAGBANK type.  This is similar to FLAG except it only accepts system flag banks.  Custom flag banks cannot be created.
- Added special handling for long vs int for NUMBER and STAT field types.  These types are only used in the methods file.  Scripting perceives everything as longs internally.
- Methods and functions now require the constant keyword to return constant values when they return reference values.
- Added more entity fields
- TODO: get foreach with STRINGS working
- TODO: make foreach with ARRAYS skip NULLs when appropriate
- TODO: add missing processing for remaining game types

+++ merc.h
@@ -7310,11 +7310,11 @@ struct exit_data
         ROOM_INDEX_DATA *to_room;
         WNUM_LOAD wnum;
     } u1;
-    int exit_info;
+    long exit_info;
     char *keyword;
     char *short_desc; /* short description for displaying only */
     char *long_desc;  /* long desc for closer scrutiny */
-    int rs_flags;
+    long rs_flags;
     int orig_door;
 
     struct door_data
